+++ linphone-app/ui/modules/Linphone/Account/AccountStatus.qml
@@ -17,6 +17,7 @@ Item {
 	signal clicked
 	property alias cursorShape:mouseArea.cursorShape
 	property alias betterIcon : presenceLevel.betterIcon
+	property bool noAccountConfigured: AccountSettingsModel.accounts.length <= ((SettingsModel.showLocalSipAccount ? 1 : 0))
 	
 	// ---------------------------------------------------------------------------
 	MouseArea {
@@ -41,7 +42,7 @@ Item {
 				Layout.bottomMargin: AccountStatusStyle.presenceLevel.bottomMargin
 				Layout.preferredHeight: AccountStatusStyle.presenceLevel.size
 				Layout.preferredWidth: AccountStatusStyle.presenceLevel.size
-				
+				visible: !accountStatus.noAccountConfigured
 				PresenceLevel {
 					id:presenceLevel
 					anchors.fill:parent
@@ -67,19 +68,24 @@ Item {
 			
 			Text {
 				id:username
-				Layout.alignment: Qt.AlignBottom | Qt.AlignLeft
+				Layout.fillWidth: true
+				Layout.alignment: accountStatus.noAccountConfigured ?  Qt.AlignVCenter | Qt.AlignLeft: Qt.AlignBottom | Qt.AlignLeft
 				color: AccountStatusStyle.username.color
 				elide: Text.ElideRight
 				font.bold: true
 				font.pointSize: AccountStatusStyle.username.pointSize
-				text: AccountSettingsModel.username
+				//: 'No account configured' : Status text when there is no configured account.
+				text: accountStatus.noAccountConfigured  ? qsTr('noAccount'): AccountSettingsModel.username
 				verticalAlignment: Text.AlignBottom
+				wrapMode: Text.WordWrap
+				maximumLineCount: 3
 			}
 			Item {
 				Layout.alignment: Qt.AlignBottom | Qt.AlignLeft
 				Layout.bottomMargin: 5
 				Layout.preferredHeight: AccountStatusStyle.presenceLevel.size
 				Layout.preferredWidth: AccountStatusStyle.presenceLevel.size
+				visible: !accountStatus.noAccountConfigured
 				MessageCounter {
 					id: messageCounter
 					anchors.fill: parent
@@ -99,6 +105,7 @@ Item {
 		Text {
 			Layout.preferredHeight:parent.height / 2
 			Layout.preferredWidth:parent.width
+			visible: !accountStatus.noAccountConfigured
 			color: AccountStatusStyle.sipAddress.color
 			elide: Text.ElideRight
 			font.pointSize: AccountStatusStyle.sipAddress.pointSize